fixed string formatting problem

+++ domoticz/hardware/ZWaveBase.cpp
@@ -188,7 +188,7 @@ void ZWaveBase::SendSwitchIfNotExists(const _tZWaveDevice *pDevice)
 		unsigned long lID = (ID1 << 24) + (ID2 << 16) + (ID3 << 8) + ID4;
 
 		char szID[10];
-		sprintf(szID, "%08x", lID);
+		sprintf(szID, "%08x", (unsigned int)lID);
 		std::string ID = szID;
 		unsigned char unitcode = 1;
